Updates and improvements to City object
Added a new function called `search_country` to the `aladhan.location_types` module. This function takes a query and searches for a country in the `aladhan.countries.countries` dictionary. It then returns a dictionary with the country's alpha-2, alpha-3 codes, and CLDR name, made for `City` object especially.

Also made some updates to the `City` object:

- It now accepts alpha-3 codes and CLDR country names in addition to alpha-2 codes.
- A validator has been added to the `City` object to ensure that the city and country are valid. If an invalid location is provided, an `InvalidLocationException` is raised.
- The City object now has a new method called `get_text()`, which returns a string in the format "CITY, STATE, COUNTRY" or "CITY, COUNTRY", depending on whether the state is provided.